21.Ders

+++ Csharp-Bootcamp-301-EntityLayer/Concrete/Product.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -11,10 +12,13 @@ namespace Csharp_Bootcamp_301_EntityLayer.Concrete
         public int ProductID { get; set; }
         public string ProductName { get; set; }
         public int ProductStock {  get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public string ProductDescription { get; set; }
         public int CategoryID { get; set; }
         public virtual Category Category { get; set; }
         public List<Order> Orders { get; set; }
+
+        //[NotMapped]
+        //public string CategoryName { get; set; }
     }
 }
